partial commit... Update ClientState and add client IRC engine
This commit includes the addition of a client IRC engine in the 'client_irc_engine' module and the expansion of the ClientState interface along with the MutableClientState data class. Main additions within these classes are 'nickState' and 'self' elements. This update aims at enhancing the interactivity of the client by equipping it with the capacity to engage with and reply to incoming IRC messages.