Add runWaitingTest function to TestUtils
The function runWaitingTest has been added to TestUtils, enabling more efficient testing. It utilizes a block of code that, when executed, completes a CompletableDeferred object. The function was previously defined in EngineTest but has been moved to TestUtils for better reusability. Its calls in EngineTest.kt have been updated accordingly.