Add PrivateChatTabView, update dependencies and refactor code
A new Composable `PrivateChatTabView` is added for creating tabs for private chats in the UI. The UI dependencies in `build.gradle.kts` are updated and organized for `commonMain`. Numerous file changes, method and variable renames, and refactorings were done to improve the maintainability and readability of the code. These changes have also allowed the application UI to follow Dependency Inversion Principle (DIP) thereby making the architecture more modular and independent of concrete implementations.